service-2/requirements.txt
django
djangorestframework
requests
grpcio
grpcio-tools
